aws-data-platform/terraform/uploads/mwaa/requirements/requirements.txt
dbt-snowflake==1.1.0
websocket-client==0.54.0
text-unidecode==1.2
apache-airflow-providers-snowflake==3.0.0
apache-airflow-providers-dbt-cloud==2.0.0
